abm con el rol premium implementado

+++ src/controllers/products.controller.js
@@ -47,8 +47,8 @@ export const findOneprod = async (req,res) => {
 }
 
 export const createOneProd = async (req, res) => {
-    const { title, description, code, category, price, stock, status, tumbnail } = req.body
-    if (!title || !description || !code || !category  || !price|| !stock|| !status || !tumbnail) {
+    const { title, description, code, category, price, stock, status, owner, tumbnail } = req.body
+    if (!title || !description || !code || !category  || !price|| !stock|| !status || !owner|| !tumbnail) {
       //return res.status(400).json({ message: 'Data missing' })
       CustomError.createError({
         name: 'Product creation error',
@@ -74,10 +74,15 @@ export const createOneProd = async (req, res) => {
 
   export const updateOneProd = async (req,res) => {
     const id = req.params.id
-    const {title, description, code, category, price, stock, status, tumbnail} = req.body
+    const {title, description, code, category, price, stock, status, owner, tumbnail} = req.body
     try {
-        const updateProd = await updateOneProduct({_id:id},{title, description, code, category, price, stock, status, tumbnail})
+        if(req.user.role === 'premium' || req.user.id !== owner){
+            res.status(401).json({message: 'Ud no puede modificar este producto'})
+        }else{
+            const updateProd = await updateOneProduct({_id:id},{title, description, code, category, price, stock, status, owner,tumbnail})
         res.status(200).json({ message: 'Product updated', product: updateProd })
+        }
+        
         
     } catch(error) {
         res.status(500).json({ error })
@@ -86,9 +91,17 @@ export const createOneProd = async (req, res) => {
 
  export const deleteProd = async (req,res) => {
     const id = req.params.id
+    const {owner} = req.body
+    console.log(owner)
     try{
-        const deleteOneProd = await deleteOneProduct({_id: id})
-        res.status(200).json({ message: 'Product deleted' })
+        if(req.user.role === 'premium' || req.user.id !== owner){
+            res.status(401).json({message: 'Ud no puede eliminar este producto'})
+        }else{
+            const deleteOneProd = await deleteOneProduct({_id: id})
+            res.status(200).json({ message: 'Product deleted' })
+        }
+        
+        
     
     }catch(error){
         res.status(500).json({ error })
